requirements.txt
GitPython
numpy
future

### todo: Check if there is a way to avoid requiring pytorch
torchnet
